requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
multitasking==0.0.9
numerize==0.12
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
slack-bolt==1.4.4
slack-sdk==3.4.2
tweepy==3.10.0
typing-extensions==3.10.0.0
urllib3==1.26.3
websocket-client==0.58.0
Werkzeug==2.0.1
whitenoise==5.2.0
yfinance==0.1.63
zipp==3.4.1
